ticket24: check if tested feature is present.

+++ testsuite/gna/ticket24/testsuite.sh
@@ -2,25 +2,32 @@
 
 . ../../testenv.sh
 
-analyze -fpsl psl.vhdl
-elab_simulate -fpsl psl --psl-report=psl.out
+GHDL_STD_FLAGS="-fpsl"
+analyze psl.vhdl
+elab psl
+if ghdl_has_feature psl psl; then
+  simulate psl --psl-report=psl.out
 
-if ! diff --strip-trailing-cr psl.out psl.ref > /dev/null; then
-    echo "report mismatch"
-    exit 1
+  if ! diff --strip-trailing-cr psl.out psl.ref > /dev/null; then
+      echo "report mismatch"
+      exit 1
+  fi
+
+  rm -f psl.out
 fi
-
-rm -f psl.out
 clean
 
 # Using vhdl 08
 GHDL_STD_FLAGS="-fpsl --std=08"
 analyze psl.vhdl
-elab_simulate psl --psl-report=psl.out
+elab -fpsl psl
+if ghdl_has_feature psl psl; then
+  simulate psl --psl-report=psl.out
 
-diff --strip-trailing-cr -q psl.out psl.ref
+  diff --strip-trailing-cr -q psl.out psl.ref
 
-rm -f psl.out
+  rm -f psl.out
+fi
 clean
 
 # Usage example (python 2.7):
